Bump the version of vim-scala-git
A commit (scala/scala-dist@5e8501a335f5cfb13d08cb5c8b40c3c2d293b45d) is added to the master branch of the upstream git repository 10 month ago.
This package was older than that. In order to let the AUR helpers upgrade the package, I have to bump the version of the package,
even though it is too late.